% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_unity_table.R
\name{apply_unity_table}
\alias{apply_unity_table}
\title{Apply Unity Table to Standardize Categories}
\usage{
apply_unity_table(unity_table, table, column)
}
\arguments{
\item{unity_table}{A data frame created by \code{make_unity_table} containing the standardized values and their variations.}

\item{table}{The data frame containing the column to be standardized.}

\item{column}{The name of the column in \code{table} to be standardized.}
}
\value{
The input data frame with the specified column standardized according to the unity table.
}
\description{
This function applies a unity table to a specified column in a data frame, uniting all variations of a value into a single standardized value.
}
\examples{
df <- data.frame(ID = 1:6, Status = c("single", "Single", "SINGLE", "married", "Married", "MARRIED"))
unity_table <- make_unity_table(df, "Status", char_range = 1)
df_standardized <- apply_unity_table(unity_table, df, "Status")
print(df_standardized)

}
