% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_unite_column.R
\name{auto_unite_column}
\alias{auto_unite_column}
\title{Automatically Unite Variations within a Column}
\usage{
auto_unite_column(table, column)
}
\arguments{
\item{table}{The data frame containing the column to be standardized.}

\item{column}{The name of the column in \code{table} to be standardized.}
}
\value{
The input data frame with the specified column contents standardized and unified.
}
\description{
This function automatically identifies and unites variations of values within a specified column in a data frame. It uses \code{make_unity_table} to create a unity table and \code{apply_unity_table} to standardize the column based on the unity table.
}
\examples{
df <- data.frame(ID = 1:6, Status = c("single", "Single", "SINGLE", "married", "Married", "MARRIED"))
df_standardized <- auto_unite_column(df, "Status")
print(df_standardized)

}
