% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_marital_status}
\alias{clean_marital_status}
\title{Marital Status Dataset}
\format{
A data frame with the following columns:
\describe{
\item{ID}{A unique identifier for each individual.}
\item{Marital Status}{The marital status of the individual, indicated as "married" or "single".}
\item{Have a Job}{Indicates whether the individual has a job, answered with "yes" or "no".}
\item{Financially Satisfied}{Indicates whether the individual is financially satisfied, answered with "yes" or "no".}
}
}
\source{
Simulated data for practicing unite functions.
}
\usage{
clean_marital_status
}
\description{
A dataset containing personal information written with some bad variation.
}
\examples{
data(clean_marital_status)
}
\keyword{datasets}
