% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_column.R
\name{find_column}
\alias{find_column}
\title{Search for a Wanted Column in a Set of Tables}
\usage{
find_column(search_in, search_for, fuzzy = TRUE)
}
\arguments{
\item{search_in}{A list of dataframes in which to search for the specified column name.}

\item{search_for}{A character string representing the name of the column to search for.}

\item{fuzzy}{A logical value indicating whether fuzzy matching should be used. Default is TRUE.}
}
\value{
A dataframe with the indices of the tables, the indices of the matched columns within the tables, and the matched column names. Returns NULL if no matches are found.
}
\description{
This function searches for a specified column name within a list of dataframes.
It provides an option for fuzzy matching, allowing for minor differences between
the specified column name and the actual column names in the dataframes.
}
\examples{
df3 <- data.frame(Name = c("Alice", "Bob"), Age = c(25, 30))
df4 <- data.frame(Name = c("Carol", "David"), Salary = c(50000, 60000))
search_list2 <- list(df3, df4)
find_column(search_list2, "name")
find_column(search_list2, "salary", fuzzy = TRUE)

}
