% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_value.R
\name{find_value}
\alias{find_value}
\title{Get the Index of All Matches in a Vector}
\usage{
find_value(target, ref_vector, fuzzy = FALSE)
}
\arguments{
\item{target}{The target string that you are looking for.}

\item{ref_vector}{The vector that will be searched.}

\item{fuzzy}{A logical value indicating whether fuzzy matching should be used. Default is FALSE.}
}
\value{
A vector of indices where the target string matches elements in the reference vector. Returns NULL if no matches are found.
}
\description{
This function searches for a target string within a reference vector and returns the indices of all matches.
It provides an option for fuzzy matching, allowing for minor differences between the target string and the elements in the reference vector.
}
\examples{
ref_vector <- c("apple", "banana", "cherry", "date")
find_value("apple", ref_vector)
find_value("banana", ref_vector, fuzzy = TRUE)

ref_vector2 <- c("Allice", "Bob", "Carol", "David")
find_value("Alice", ref_vector2, fuzzy = TRUE)
find_value("Dave", ref_vector2, fuzzy = TRUE)

}
