% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_match.R
\name{first_match}
\alias{first_match}
\title{Get the Index of the First Match in a Vector}
\usage{
first_match(target, ref_vector, fuzzy = FALSE)
}
\arguments{
\item{target}{The target string that you are looking for.}

\item{ref_vector}{The vector that will be searched.}

\item{fuzzy}{A logical value indicating whether fuzzy matching should be used. Default is FALSE.}
}
\value{
The index of the first match where the target string matches an element in the reference vector. Returns NULL if no match is found.
}
\description{
This function searches for a target string within a reference vector and returns the index of the first match.
It provides an option for fuzzy matching, allowing for minor differences between the target string and the elements in the reference vector.
}
\examples{
ref_vector <- c("apple", "banana", "cherry", "date")
first_match("apple", ref_vector)
first_match("banana", ref_vector, fuzzy = TRUE)

ref_vector2 <- c("Allice", "Bob", "Carol", "David")
first_match("Alice", ref_vector2, fuzzy = TRUE)
first_match("Dave", ref_vector2, fuzzy = TRUE)

}
