% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unity_table.R
\name{make_unity_table}
\alias{make_unity_table}
\title{Create a "Unity Table" for Categorical Column}
\usage{
make_unity_table(table, column, char_range = 1)
}
\arguments{
\item{table}{A data frame containing the column to be processed.}

\item{column}{The name of the categorical column to be standardized.}

\item{char_range}{Numeric value indicating the character range for fuzzy matching (default is 1).}
}
\value{
A data frame with two columns: \code{value} containing the standardized category names, and \code{variations} containing lists of the original variations.
}
\description{
This function reads all the values in a categorical column and recognizes all the variations in the writing of each unique category name. It unites these variations into one value to cure categorical inconsistency.
}
\examples{
df <- data.frame(ID = 1:6, Status = c("single", "Single", "SINGLE", "married", "Married", "MARRIED"))
unity_table <- make_unity_table(df, "Status", char_range = 1)
print(unity_table)

}
