% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_bill}
\alias{match_bill}
\title{Bill Dataset}
\format{
A data frame with the following column:
\describe{
\item{Product ID}{A unique identifier for each vegetable product.}
\item{Product Name}{The name of the vegetable product.}
\item{Cost}{The cost of the vegetable product.}
}
}
\source{
Simulated data for practice with matching and standardize functions.
}
\usage{
match_bill
}
\description{
A dataset containing product IDs from a bill; You need to match each ID with its Name an Cost from match_vegetables, match_fruits, and match_missings.
}
\examples{
data(match_bill)
}
\keyword{datasets}
