% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_fruits}
\alias{match_fruits}
\title{Fruit Products Dataset}
\format{
A data frame with the following columns:
\describe{
\item{Product ID}{A unique identifier for each fruit product.}
\item{Product Name}{The name of the fruit product.}
\item{Cost}{The cost of the fruit product.}
}
}
\source{
Simulated data for practice with matching and standardize functions.
}
\usage{
match_fruits
}
\description{
A dataset containing product IDs, names, and costs for fruits; needs to be matched with match_bill.
}
\examples{
data(match_fruits)
}
\keyword{datasets}
