% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_missings}
\alias{match_missings}
\title{Missing Products Dataset}
\format{
A data frame with the following columns:
\describe{
\item{Product ID}{A unique identifier for each missing product.}
\item{Product Name}{The name of the missing product.}
\item{Cost}{The cost of the missing product.}
}
}
\source{
Simulated data for practice with matching and standardize functions.
}
\usage{
match_missings
}
\description{
A dataset containing product IDs, names, and costs for products missing from the vegetable and fruit datasets; needs to be matched with match_bill.
}
\examples{
data(match_missings)
}
\keyword{datasets}
