% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_backbone.R
\name{match_string_to_string}
\alias{match_string_to_string}
\title{An Efficient Algorithm for Fuzzy Search}
\usage{
match_string_to_string(wanted_word, ref_word, fuzzy = TRUE, char_range = 2)
}
\arguments{
\item{wanted_word}{A character string representing the first word to be matched.}

\item{ref_word}{A character string representing the reference word to be compared.}

\item{fuzzy}{A logical value indicating whether fuzzy matching should be used. Default is TRUE.}

\item{char_range}{An integer specifying the maximum tolerated number of added or removed characters in the reference word. Default is 2.}
}
\value{
A numeric value representing the match accuracy between the two words, or NULL if the words do not match.
}
\description{
This function matches two words and calculates their match accuracy.
It provides an option for fuzzy matching, allowing for minor differences
between the words.
}
\examples{
match_string_to_string("example", "exampel")
match_string_to_string("test", "tst", fuzzy = TRUE, char_range = 1)

}
