% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_tables.R
\name{match_tables}
\alias{match_tables}
\title{Match Multiple Tables and Gather the Data in One Place}
\usage{
match_tables(get_from, add_to, get_columns, according_to, fuzzy = FALSE)
}
\arguments{
\item{get_from}{A list of dataframes containing the tables that will provide data.}

\item{add_to}{A dataframe that will receive the new data for each match.}

\item{get_columns}{A vector of column names containing the data needed to be imported for each match.}

\item{according_to}{A mutual column between the tables, which the matching process is based on.}

\item{fuzzy}{A logical value indicating whether fuzzy matching should be used. Default is FALSE.}
}
\value{
A dataframe with the data from the specified columns added to the \code{add_to} dataframe.
}
\description{
This function matches multiple tables based on a mutual column and gathers specified columns of data into a single dataframe.
It provides an option for fuzzy matching, allowing for minor differences between the values in the mutual column.
}
\examples{
df4 <- data.frame(Name = c("Alice", "Bob"), Age = c(25, 30))
df5 <- data.frame(Name = c("Carol", "David", "Allice"), Salary = c(40000, 50000, 60000))
get_from_list2 <- list(df4, df5)
add_to_df2 <- data.frame(Name = c("Alice", "Bob", "Carol", "David"))
match_tables(get_from_list2, add_to_df2, c("Age", "Salary"), "Name", fuzzy = TRUE)

}
