% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_two_tables.R
\name{match_two_tables}
\alias{match_two_tables}
\title{Match Two Tables and Gather the Data in One of Them}
\usage{
match_two_tables(get_from, add_to, get_column, according_to, fuzzy = FALSE)
}
\arguments{
\item{get_from}{A dataframe that will provide data.}

\item{add_to}{A dataframe that will receive the new data for each match.}

\item{get_column}{The column containing the data needed to be imported for each match.}

\item{according_to}{A mutual column between the tables, which the matching process is based on.}

\item{fuzzy}{A logical value indicating whether fuzzy matching should be used. Default is FALSE.}
}
\value{
A dataframe with the data from the specified column added to the \code{add_to} dataframe.
}
\description{
This function matches two tables based on a mutual column and gathers the specified column of data from one table into another.
It provides an option for fuzzy matching, allowing for minor differences between the values in the mutual column.
}
\examples{
df3 <- data.frame(Name = c("Alice", "Bob"), Age = c(25, 30))
df4 <- data.frame(Name = c("Alice", "Bob"), Salary = c(50000, 60000))
match_two_tables(df3, df4, "Age", "Name", fuzzy = TRUE)

}
