% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_vegetables}
\alias{match_vegetables}
\title{Vegetable Products Dataset}
\format{
A data frame with the following columns:
\describe{
\item{Product ID}{A unique identifier for each vegetable product.}
\item{Name}{The name of the vegetable product.}
\item{Cost}{The cost of the vegetable product.}
}
}
\source{
Simulated data for practice with matching and standardize functions.
}
\usage{
match_vegetables
}
\description{
A dataset containing product IDs, names, and costs for vegetables; needs to be matched with match_bill.
}
\examples{
data(match_vegetables)
}
\keyword{datasets}
