% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_backbone.R
\name{split_and_remove_special}
\alias{split_and_remove_special}
\title{Split String into Letters and Remove Special Characters}
\usage{
split_and_remove_special(word)
}
\arguments{
\item{word}{A character string that will be split into letters and cleaned from special characters.}
}
\value{
A character vector of letters from the input word, with special characters removed.
}
\description{
This helper function takes a word as input, splits it into individual letters,
and removes any special characters from the resulting list of letters.
}
\examples{
split_and_remove_special("Hello! How's it going?")
split_and_remove_special("Test@123")

}
