% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Convert Text to Standardized Version}
\usage{
standardize(word, lowercase = TRUE)
}
\arguments{
\item{word}{The string that will be standardized.}

\item{lowercase}{If TRUE, the output will be in lowercase. Default is TRUE.}
}
\value{
A standardized version of the input string.
}
\description{
This function converts a given string into a standardized version by removing unwanted characters and optionally converting it to lowercase.
}
\examples{
standardize("Hello, World!")
standardize("Special @# Characters!", lowercase = FALSE)

}
