% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_col_names.R
\name{standardize_col_names}
\alias{standardize_col_names}
\title{Standardize the Column Names of a Given Table}
\usage{
standardize_col_names(table, lowercase = TRUE)
}
\arguments{
\item{table}{The table whose columns will be standardized. Must be a data frame.}

\item{lowercase}{If TRUE, the column names will be converted to lowercase. Default is TRUE.}
}
\value{
The table with standardized column names.
}
\description{
This function standardizes the column names of a given table so that other functions can work properly.
The standardization process can include converting column names to lowercase if specified.
}
\examples{
df <- data.frame("First Name." = c("Alice", "Bob"), "Age" = c(25, 30))
standardize_col_names(df, lowercase = TRUE)

}
