% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_vector.R
\name{standardize_vector}
\alias{standardize_vector}
\title{Standardize Values Within a Vector}
\usage{
standardize_vector(values, lowercase = TRUE)
}
\arguments{
\item{values}{The vector of values to be standardized.}

\item{lowercase}{If TRUE, the standardized values will be converted to lowercase. Default is TRUE.}
}
\value{
A vector of standardized values.
}
\description{
This function standardizes the values within a vector. It processes each value by applying a standardization function, and can convert the values to lowercase if specified.
}
\examples{
values <- c("Hello", "World", "FOO", "Bar Chart")
standardize_vector(values, lowercase = TRUE)

}
