% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_values_in_table.R
\name{unite_values_in_table}
\alias{unite_values_in_table}
\title{Unite Variations of Strings in a Table}
\usage{
unite_values_in_table(
  table,
  target_values,
  unite_to,
  fuzzy = TRUE,
  char_range = 1
)
}
\arguments{
\item{table}{The dataframe in which variations of strings will be united.}

\item{target_values}{A vector of variations that will be united. If \code{fuzzy} is TRUE, you can provide a smaller set of variations as the function will use fuzzy matching to recognize similar strings.}

\item{unite_to}{The final standardized form to which all variations will be united.}

\item{fuzzy}{When TRUE, fuzzy matching is used to recognize variations. Default is TRUE.}

\item{char_range}{The maximum tolerated number of added or removed characters for fuzzy matching. Default is 1.}
}
\value{
The dataframe with variations of strings united according to the specified parameters.
}
\description{
This function unites all variations of specific strings within a table to avoid minor misspellings and ensure consistency for statistical analysis. Variations can be united using fuzzy matching if enabled.
}
\examples{
# Sample dataframe
df <- data.frame(Name = c("Jonh Doe", "John Doe", "Jon Doe", "Jane Doe"))

# Unite variations of "John Doe" into a single form
df_united <- unite_values_in_table(df, c("Jonh Doe"), "John Doe", fuzzy = TRUE)

}
